Move vaults table to --lib
